test(psu): update Keysight E36100 tests
